removeSupplierOffer changes

+++ evolution/src/main/java/com/evolving/nglm/evolution/GUIManager.java
@@ -23040,9 +23040,7 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot) thro
                 OfferVoucher voucher = (OfferVoucher) OfferProductVoucherAndSupplierIDs.get("offerVoucher");                
                 
                 Offer offer = (Offer) offerObject;
-                String offerName = offer.getGUIManagedObjectName();
-                
-               
+                String offerName = offer.getGUIManagedObjectName();                               
                 if (product != null)
                   {
                     String productID = product.getProductID();
@@ -23068,7 +23066,7 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot) thro
                     }
                   }
                 if (voucher != null)
-                  {
+                  {nn
                     String voucherID = voucher.getVoucherID();
                     String voucherName = (voucherService.getStoredVoucher(voucherID)).getGUIManagedObjectName();
 
@@ -23195,6 +23193,7 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot) thro
         else if (activeSupplier != null && activeSupplier != "InactiveReseller")
           {
             Offer offer = (Offer) offerObject;
+            String offerName = offer.getGUIManagedObjectName();
             Map<String, Object> OfferProductVoucherAndSupplierIDs = OfferProductVoucherAndSupplierIDs(
                 (Offer) offerObject);
             String supplierID = (String) OfferProductVoucherAndSupplierIDs.get("supplierID");
@@ -23203,9 +23202,11 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot) thro
             
             if (product != null)
               {
-                if (activeSupplier.equals(supplierID))
+                String productID = product.getProductID();
+                String productName = (productService.getStoredProduct(productID)).getGUIManagedObjectName();
+                if (activeSupplier.equals(supplierID) && offerName.equals(productName))
                   {
-                    String productId = product.getProductID();                  
+                    String productId = product.getProductID();
                     productService.removeProduct(productId, userID);
                     offerService.removeOffer(offerID, userID);
 
@@ -23214,7 +23215,8 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot) thro
                     response.put("responseCode", responseCode);
                   }
 
-                else
+                else if (offerName.equals(productName) && !(activeSupplier.equals(supplierID)))
+
                   {
 
                     response.put("responseCode",
@@ -23224,11 +23226,18 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot) thro
                     return JSONUtilities.encodeObject(response);
 
                   }
+                else
+                  {
+                    if (log.isInfoEnabled())
+                      log.info("offer is not supplierOffer");
+                  }
               }
             if (voucher != null)
               {
+                String voucherID = voucher.getVoucherID();
+                String voucherName = (voucherService.getStoredVoucher(voucherID)).getGUIManagedObjectName();
 
-                if (activeSupplier.equals(supplierID))
+                if (activeSupplier.equals(supplierID) && offerName.equals(voucherName))
                   {
                     String voucherId = voucher.getVoucherID();
                     voucherService.removeVoucher(voucherId, userID, uploadedFileService);
@@ -23239,7 +23248,8 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot) thro
                     response.put("responseCode", responseCode);
                   }
 
-                else
+                else if (offerName.equals(voucherName) && !(activeSupplier.equals(supplierID)))
+
                   {
 
                     response.put("responseCode",
@@ -23249,6 +23259,11 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot) thro
                     return JSONUtilities.encodeObject(response);
 
                   }
+              }
+            else
+              {
+                if (log.isInfoEnabled())
+                  log.info("offer is not supplierOffer");
               }
 
           }
